Fix running golangci-lint in different packages
The built-in typecheck cannot run on single files and will cause a
compilation error as soon as we attempt to lint a file that requires a
file from the same package. As a best-effort workaround now linting
the entire package a file has been changed in in the pre-commit hook.

https://github.com/golangci/golangci-lint/issues/1574#issuecomment-804500358
https://github.com/golangci/golangci-lint/issues/2912